Fix jsvc compilation warnings

+++ src/native/unix/native/jsvc-unix.c
@@ -589,14 +589,21 @@ retry:
         return -1;
     }
     else {
-        lockf(fd, F_LOCK, 0);
+        if (lockf(fd, F_LOCK, 0)) {
+            log_error("check_pid: Failed to lock PID file [%s] with file descriptor [%d] for reading due to [%d]",
+                    args->pidf, fd, errno);
+            return -1;
+        }
         i = read(fd, buff, sizeof(buff));
         if (i > 0) {
             buff[i] = '\0';
             pid = atoi(buff);
             if (kill(pid, 0) == 0) {
                 log_error("Still running according to PID file %s, PID is %d", args->pidf, pid);
-                lockf(fd, F_ULOCK, 0);
+                if (lockf(fd, F_ULOCK, 0)) {
+                    log_error("check_pid: Failed to unlock PID file [%s] with file descriptor [%d] after reading due to [%d]",
+                            args->pidf, fd, errno);
+                }
                 close(fd);
                 return 122;
             }
@@ -606,7 +613,10 @@ retry:
         fprintf(pidf, "%d\n", (int)getpid());
         fflush(pidf);
         fclose(pidf);
-        lockf(fd, F_ULOCK, 0);
+        if (lockf(fd, F_ULOCK, 0)) {
+            log_error("check_pid: Failed to unlock PID file [%s] with file descriptor [%d] after reading due to [%d]",
+                    args->pidf, fd, errno);
+        }
         close(fd);
     }
     return 0;
@@ -625,7 +635,11 @@ static void remove_pid_file(arg_data *args, int pidn)
     if (fd < 0) {
         return;
     }
-    lockf(fd, F_LOCK, 0);
+    if (lockf(fd, F_LOCK, 0)) {
+        log_error("remove_pid_file: Failed to lock PID file [%s] with file descriptor [%d] for reading due to [%d]",
+                args->pidf, fd, errno);
+        return;
+    }
     i = read(fd, buff, sizeof(buff));
     if (i > 0) {
         buff[i] = '\0';
@@ -643,7 +657,10 @@ static void remove_pid_file(arg_data *args, int pidn)
             ("remove_pid_file: pid changed (%d->%d), not removing pid file %s",
              pidn, pid, args->pidf);
     }
-    lockf(fd, F_ULOCK, 0);
+    if (lockf(fd, F_ULOCK, 0)) {
+        log_error("remove_pid_file: Failed to unlock PID file [%s] with file descriptor [%d] after reading due to [%d]",
+                args->pidf, fd, errno);
+    }
     close(fd);
 }
 
@@ -663,9 +680,16 @@ static int get_pidf(arg_data *args, bool quiet)
         /* something has gone wrong the JVM has stopped */
         return -1;
     }
-    lockf(fd, F_LOCK, 0);
+    if (lockf(fd, F_LOCK, 0)) {
+        log_error("get_pidf: Failed to lock PID file [%s] with file descriptor [%d] for reading due to [%d]",
+                args->pidf, fd, errno);
+        return -1;
+    }
     i = read(fd, buff, sizeof(buff));
-    lockf(fd, F_ULOCK, 0);
+    if (lockf(fd, F_ULOCK, 0)) {
+        log_error("get_pidf: Failed to unlock PID file [%s] with file descriptor [%d] after reading due to [%d]",
+                args->pidf, fd, errno);
+    }
     close(fd);
     if (i > 0) {
         buff[i] = '\0';
@@ -759,9 +783,16 @@ static int wait_child(arg_data *args, int pid)
             /* something has gone wrong the JVM has stopped */
             return 1;
         }
-        lockf(fd, F_LOCK, 0);
+        if (lockf(fd, F_LOCK, 0)) {
+            log_error("wait_child: Failed to lock PID file [%s] with file descriptor [%d] for reading due to [%d]",
+                    args->pidf, fd, errno);
+            return 1;
+        }
         i = read(fd, buff, sizeof(buff));
-        lockf(fd, F_ULOCK, 0);
+        if (lockf(fd, F_ULOCK, 0)) {
+            log_error("wait_child: Failed to unlock PID file [%s] with file descriptor [%d] after reading due to [%d]",
+                    args->pidf, fd, errno);
+        }
         close(fd);
         if (i > 0) {
             buff[i] = '\0';
@@ -1061,7 +1092,9 @@ static void set_output(char *outfile, char *errfile, bool redirectstdin, char *p
     int fork_needed = 0;
 
     if (redirectstdin == true) {
-        freopen("/dev/null", "r", stdin);
+        if (freopen("/dev/null", "r", stdin) == NULL) {
+            log_error("Failed to redirect stdin to /dev/null due to [%d]", errno);
+        }
     }
 
     log_debug("redirecting stdout to %s and stderr to %s", outfile, errfile);
@@ -1072,7 +1105,9 @@ static void set_output(char *outfile, char *errfile, bool redirectstdin, char *p
     if (strcmp(outfile, "&1") == 0 && strcmp(errfile, "&2") == 0)
         return;
     if (strcmp(outfile, "SYSLOG") == 0) {
-        freopen("/dev/null", "a", stdout);
+        if (freopen("/dev/null", "a", stdout) == NULL) {
+            log_error("Failed to redirect stdout to /dev/null due to [%d]", errno);
+        }
         /* Send stdout to syslog through a logger process */
         if (pipe(out_pipe) == -1) {
             log_error("cannot create stdout pipe: %s", strerror(errno));
@@ -1090,7 +1125,9 @@ static void set_output(char *outfile, char *errfile, bool redirectstdin, char *p
     }
 
     if (strcmp(errfile, "SYSLOG") == 0) {
-        freopen("/dev/null", "a", stderr);
+        if (freopen("/dev/null", "a", stderr) == NULL) {
+            log_error("Failed to redirect stderr to /dev/null due to [%d]", errno);
+        }
         /* Send stderr to syslog through a logger process */
         if (pipe(err_pipe) == -1) {
             log_error("cannot create stderr pipe: %s", strerror(errno));
